pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurora"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
    "hypothesis>=6.0.0",
    "coverage[toml]>=7.0.0",
    "vulture>=2.0.0",
    "deptry>=0.20.0",
    "bandit>=1.8.0",
    "pip-audit>=2.0.0",
    "scalene>=1.5.0",
    "pytest-benchmark>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --maxfail=1 --disable-warnings"
markers = [
    "smoke: deterministic minimum-path tests",
    "quarantine: unstable/flaky; excluded from CI by default"
]

[tool.coverage.run]
source = ["src/aurora"]
branch = true

[tool.coverage.report]
skip_covered = true
show_missing = true
fail_under = 80

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]
# Start permissive, then ratchet up
select = ["E","F","I"]     # errors, flakes, import order
ignore = ["E501"]          # wrap later

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]
packages = ["aurora"]
strict = true
# Temporarily allow 3rd-party stubs gaps:
ignore_missing_imports = true

[tool.bandit]
skips = ["B101"]           # assert used for invariants; review later